+++ changelog.md
@@ -4,6 +4,8 @@
 * Fixed mobs attack speed being doubled
   * Mojang's melee attack goal uses a method that halves the attack speed cooldown ... for no reason
 * Mobs max attack speed is now 2 instead of 1.25
+* Decreased skeleton's shooting cooldown
+  * Also, hard mode no longer makes skeletons shoot faster
 
 ## 2.5.2
 * Witches, Skeletons and Pillagers now actually flee from the target instead of running away from any Player